Update OAuthSampleApp.java
Changes in the client application done as per comments and discussion. 

SetUpDbForOAuth() -> Adding authentication record in Vertica database is having issues. It's dropping the user and authentication method, but while setting the Authentication record, there is some error because of which Token authentication is failing and Database connection with access token is unable to establish. 

Without SetUpDbForOAuth() method call, Application is woking good.